requirements.txt
numpy
pandas
joblib
scikit-learn==1.7.2
scipy
nltk
textstat
google-genai
gradio
